requirements.txt
# basic django and backend
django>=3.0.6,<4.0.0
django-decorator-include>=2.1,<3.0.0
python-memcached>=1.59,<2.0.0
psycopg2-binary>=2.8.5,<3.0.0
git+git://github.com/cfblaeb/django-sendfile.git

# excel export
openpyxl>=3.0.3,<4.0.0

# server
uwsgi>=2.0.18,<3.0.0

# django reversions
django-reversion>=3.0.7,<4.0.0
django-reversion-compare>=0.12.2,<2.0.0

# frontend
django-autocomplete-light>=3.5.1,<4.0.0
django-widgets-improved>=1.5.0,<2.0.0

# API access
djangorestframework>=3.11.0,<4.0.0
drf-yasg>=1.17.1,<2.0.0
markdown>=3.2.1,<4.0.0
django-filter>=2.2.0,<3.0.0
django-url-filter>=0.3.15,<2.0.0
django-crispy-forms>=1.9.0,<2.0.0
